src/main/java/com/agl/tenpo/application/controllers/PercentageController.java
package com.agl.tenpo.application.controllers;

import com.agl.tenpo.application.services.PercentageService;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.responses.ApiResponses;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import java.util.Map;

@RestController
@RequestMapping("api/v1/percentage")
public class PercentageController {

    private final PercentageService percentageService;

    public PercentageController(PercentageService percentageService) {
        this.percentageService = percentageService;
    }

    @Operation(summary = "Retrieve a random percentage")
    @ApiResponses(value = {
            @ApiResponse(responseCode = "200",
                    description = "Return a integer value",
                    content = {@Content(mediaType = "application/json",
                            schema = @Schema(implementation = ResponseEntity.class))}),
            @ApiResponse(responseCode = "500",
                    description = "Internal Server Error",
                    content = {@Content(mediaType = "application/json",
                            schema = @Schema(implementation = Map.class))})
    })
    @GetMapping("/")
    public ResponseEntity<Integer> retrievePercentage(){
        return new ResponseEntity<>(percentageService.retrievePercentage(), HttpStatus.OK);
    }

}
